- flyer auditor is done for this version khalas

+++ lib/b_views/z_components/dialogs/nav_dialog/nav_dialog.dart
@@ -24,7 +24,7 @@ class NavDialog extends StatelessWidget {
     @required String firstLine,
     String secondLine,
     Color color = Colorz.black255,
-    double seconds,
+    int seconds,
   }) {
 
     final Color _color = color ?? Colorz.darkRed255;
